feat (comments-service): fix

+++ backend/comments-service/src/main/java/com/github/spjavaind300/commentsservice/service/CommentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.github.spjavaind300.commentsservice.service;
 
 import com.github.spjavaind300.commentsservice.cache.ProfileCacheService;
+import com.github.spjavaind300.commentsservice.exception.UnauthorizedException;
 import com.github.spjavaind300.commentsservice.feing.AdsFeignClientInternal;
 import com.github.spjavaind300.commentsservice.dto.CommentDto;
 import com.github.spjavaind300.commentsservice.dto.ProfileDto;
@@ -12,7 +13,6 @@ import com.github.spjavaind300.commentsservice.exception.NotFoundException;
 import com.github.spjavaind300.commentsservice.mapper.CommentMapper;
 import com.github.spjavaind300.commentsservice.model.Comment;
 import com.github.spjavaind300.commentsservice.repository.CommentRepository;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -30,16 +30,11 @@ public class CommentServiceImpl implements CommentService {
     private final CommentMapper commentMapper;
     private final AdsFeignClientInternal adsFeignClientInternal;
     private final JwtUtils jwtUtils;
-    private final HttpServletRequest request;
     private final ProfileCacheService profileCacheService;
 
     @Override
     public List<CommentDto> getCommentsForAd(int adId) {
         List<Comment> comments = commentRepository.findByAdId(adId);
-
-        if (comments.isEmpty()) {
-            throw new NotFoundException("Объявление", adId);
-        }
 
         return comments.stream()
                 .map(comment -> {
@@ -88,15 +83,16 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public void deleteComment(int adId, int commentId) {
         Comment comment = commentRepository.findByIdAndAdId(commentId, adId)
-                .orElseThrow(() -> new NotFoundException("Комментарий", commentId));
+                .orElse(null);
 
-        validateCommentAccessRights(comment);
-
-        commentRepository.delete(comment);
+        if (comment != null) {
+            validateCommentAccessRights(comment);
+            commentRepository.delete(comment);
+        }
     }
 
     private void validateCommentAccessRights(Comment comment) {
-        UserContext currentUser = jwtUtils.getUserContext(request);
+        UserContext currentUser = jwtUtils.getUserContext();
 
         if (currentUser.getRole() == Role.ADMIN || currentUser.getRole() == Role.SERVICE) {
             return;
@@ -113,9 +109,9 @@ public class CommentServiceImpl implements CommentService {
     }
 
     private ProfileDto getCurrentProfile() {
-        UserContext userContext = jwtUtils.getUserContext(request);
+        UserContext userContext = jwtUtils.getUserContext();
         if (userContext == null) {
-            throw new NotFoundException("Пользователь", "не найден в контексте");
+            throw new UnauthorizedException();
         }
 
         long authorId = userContext.getAuthorId();
